[pilot] Moving clusterregistry to kube
This is just a bit of cleanup to clarify that the clusterregistry
is really a k8s-specific component.
