+++ codecov.requirement
@@ -151,7 +151,7 @@ istio.io/istio/pilot/cmd/pilot-discovery:0 [0.0]
 istio.io/istio/pilot/cmd/sidecar-injector:0 [0.0]
 istio.io/istio/pilot/pkg/bootstrap:52 [54.2]
 istio.io/istio/pilot/pkg/config/aggregate:79 [81.5]
-istio.io/istio/pilot/pkg/config/clusterregistry:73 [75.9]
+istio.io/istio/pilot/pkg/config/kube/clusterregistry:73 [75.9]
 istio.io/istio/pilot/pkg/config/kube/crd:70 [72.0]
 istio.io/istio/pilot/pkg/config/kube/ingress:94 [96.5]
 istio.io/istio/pilot/pkg/config/memory:94 [96.4]
