+++ pilot/pkg/config/kube/clusterregistry/clusterconfig.gotmpl
