+++ pilot/pkg/config/kube/clusterregistry/clusterregistry.gotmpl
