+++ pilot/pkg/config/kube/clusterregistry/conversion.go
