+++ pilot/pkg/config/kube/clusterregistry/conversion_test.go
